validation for image added

+++ list.php
@@ -1,7 +1,7 @@
 <?php 
    /*
       @Author : Mfsi_Annapurnaa
-      @purpose : handle thelisting of employee data. 
+      @purpose : handle thelisting of employee data.
                : Deletion of a row
    */
    ini_set('display_errors', 1);
@@ -13,15 +13,17 @@
    // Delete a row
    if (isset($_GET['delete']))
     {
-       
+
         $empId = $_GET['delete'];
-        
+
         // Extract image name and delete it
         $condition = ['column' => 'Employee.empId', 'operator' => '=', 'val' => $empId];
         $result = $obj->select('Employee', 'image', $condition);
         $img = mysqli_fetch_array($result);
-        unlink(IMAGEPATH . $img['image']);
-        
+        if(file_exists($img))
+        {
+            unlink(IMAGEPATH . $img['image']);
+        }
         // Delete the address detail
         $condition = ['column' => 'Address.empId', 'operator' => '=', 'val' => $empId];
         $obj->delete('Address', $condition);
@@ -29,7 +31,6 @@
         // Delete the employee detail
         $condition = ['column' => 'Employee.empId', 'operator' => '=', 'val' => $empId];
         $obj->delete('Employee', $condition);
-        
     }
 
 // Call the required query function
